+++ dj_pocs/urls.py
@@ -3,9 +3,12 @@
 from django.urls import path
 
 from .views import map_and_postcode_analysis
+from .views import admin_only_display
+
 
 app_name = "dj_pocs"
 
 urlpatterns = [
+    path("admin_view", admin_only_display, name="admin_only_display"),
     path("map", map_and_postcode_analysis, name="map_and_postcode_analysis"),
 ]